web converter commit

+++ lab11/opera/web/src/main/java/ro/ubb/opera/web/converter/BaseConverter.java
@@ -7,9 +7,9 @@ import java.util.Collection;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public abstract class BaseConverter<Model extends BaseEntity<Long>, Dto extends BaseDto>
+public abstract class BaseConverter<Model extends BaseEntity<Integer>, Dto extends BaseDto>
         implements Converter<Model, Dto> {
-    public Set<Long> convertModelsToIDs(Set<Model> models) {
+    public Set<Integer> convertModelsToIDs(Set<Model> models) {
         return models.stream()
                 .map(model -> model.getId())
                 .collect(Collectors.toSet());
